+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Improve efficiency of vector access functions: `A.Vector.at/2`,
   `A.Vector.fetch/2`, `A.Vector.replace_at/3`,  `A.Vector.delete_at/2`...  ⚡️
+* Improve compile times
 
 ### Bug fixes